# frontend/src/api/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "EvaluationDueDate": {
        "additionalProperties": false,
        "description": "``PUT`` body: the complete new value (``null`` clears the due date).",
        "properties": {
          "due_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "New due date, or null for none.",
            "title": "Due At"
          }
        },
        "required": [
          "due_at"
        ],
        "title": "EvaluationDueDate",
        "type": "object"
      },
...
      "IdeaUpdate": {
        "additionalProperties": false,
        "description": "Partial update: omitted or null fields are unchanged; ``tags`` replaces the set.",
        "properties": {
          "description_md": {
            "anyOf": [
              {
                "maxLength": 50000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description Md"
          },
          "summary": {
            "anyOf": [
              {
                "maxLength": 500,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "maxLength": 32,
                  "minLength": 1,
                  "pattern": "^[^,\\r\\n\\t]+$",
                  "type": "string"
                },
                "maxItems": 10,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "IdeaUpdate",
        "type": "object"
      },
...
      "Problem": {
        "description": "RFC 9457 problem details, returned as ``application/problem+json``.\n\nClients should branch on ``code`` (stable, snake_case), not on ``title`` or\n``detail`` (human-readable, may change). As RFC 9457 allows, ``detail``,\n``instance`` and ``request_id`` may be absent: the one exception to \"every\ndocumented response field is always present\".",
        "examples": [
          {
            "code": "not_found",
            "detail": "Idea not found.",
            "instance": "/api/v1/ideas/0b7c7d1e-7a55-4a4f-9b8b-0d7d3a9d1c11",
            "request_id": "7f9c2b4e1d6a4c0f8e3b5a2d9c1e0f7a",
            "status": 404,
            "title": "Not Found",
            "type": "urn:soundings:problem:not_found"
          }
        ],
        "properties": {
          "code": {
            "description": "Stable machine-readable error code (snake_case).",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Explanation of this occurrence.",
            "title": "Detail"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Request path that failed.",
            "title": "Instance"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Correlates with server logs (X-Request-ID).",
            "title": "Request Id"
          },
          "status": {
            "description": "HTTP status code.",
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "description": "Short, human-readable summary of the problem type.",
            "title": "Title",
            "type": "string"
          },
          "type": {
            "description": "URI identifying the problem type.",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "title": "Problem",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "session": {
        "description": "Server-side session cookie set by sign-in (Phase 1: POST /api/v1/auth/dev/login). Unsafe methods must also send the X-CSRF-Token header equal to the soundings_csrf cookie.",
        "in": "cookie",
        "name": "soundings_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Soundings REST API.\n\n* Errors are [RFC 9457](https://www.rfc-editor.org/rfc/rfc9457) problem details\n  (`application/problem+json`); branch on the `code` field.\n* Lists use cursor pagination: pass `next_cursor` back as `?cursor=`.\n* JSON fields are snake_case; timestamps are ISO 8601 with a UTC offset.\n",
    "title": "Soundings API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/ideas/{idea}": {
      "delete": {
        "description": "Project admins. Permanent, with its evaluations and comments (the SPA confirms first; there is no undo).",
        "operationId": "delete_idea",
        "parameters": [
          {
            "description": "The idea: its id (UUID) or its key such as \"CUST-12\" (any case).",
            "in": "path",
            "name": "idea",
            "required": true,
            "schema": {
              "description": "The idea: its id (UUID) or its key such as \"CUST-12\" (any case).",
              "maxLength": 36,
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}|[A-Za-z][A-Za-z0-9]{1,5}-[1-9][0-9]{0,8})$",
              "title": "Idea",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not signed in (unauthorized)"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not allowed (forbidden, csrf_failed, or a specific code)"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found, or not visible to you (not_found)"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflicts with the current state (see code)"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            },
            "description": "Validation Failed"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Delete an idea",
        "tags": [
          "ideas"
        ]
      },
      "get": {
        "operationId": "get_idea",
        "parameters": [
          {
            "description": "The idea: its id (UUID) or its key such as \"CUST-12\" (any case).",
            "in": "path",
            "name": "idea",
            "required": true,
            "schema": {
              "description": "The idea: its id (UUID) or its key such as \"CUST-12\" (any case).",
              "maxLength": 36,
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}|[A-Za-z][A-Za-z0-9]{1,5}-[1-9][0-9]{0,8})$",
              "title": "Idea",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdeaDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not signed in (unauthorized)"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found, or not visible to you (not_found)"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            },
            "description": "Validation Failed"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Get an idea (the idea page)",
        "tags": [
          "ideas"
        ]
      },
      "patch": {
        "description": "Submitter while the idea is new (else 409 idea_not_new), owner while it is not closed (409 idea_closed), project admins always. Omitted or null fields are unchanged.",
        "operationId": "update_idea",
        "parameters": [
          {
            "description": "The idea: its id (UUID) or its key such as \"CUST-12\" (any case).",
            "in": "path",
            "name": "idea",
            "required": true,
            "schema": {
              "description": "The idea: its id (UUID) or its key such as \"CUST-12\" (any case).",
              "maxLength": 36,
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}|[A-Za-z][A-Za-z0-9]{1,5}-[1-9][0-9]{0,8})$",
              "title": "Idea",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IdeaUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdeaDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not signed in (unauthorized)"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not allowed (forbidden, csrf_failed, or a specific code)"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found, or not visible to you (not_found)"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflicts with the current state (see code)"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            },
            "description": "Validation Failed"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Edit an idea",
        "tags": [
          "ideas"
        ]
      }
    },
...
    "/api/v1/ideas/{idea}/evaluation/close": {
      "post": {
        "description": "Owner or project admin. Evaluations can no longer be saved; drafts stay out of the aggregate and their authors stay blind. Idempotent. 409 idea_closed.",
        "operationId": "close_evaluation",
        "parameters": [
          {
            "description": "The idea: its id (UUID) or its key such as \"CUST-12\" (any case).",
            "in": "path",
            "name": "idea",
            "required": true,
            "schema": {
              "description": "The idea: its id (UUID) or its key such as \"CUST-12\" (any case).",
              "maxLength": 36,
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}|[A-Za-z][A-Za-z0-9]{1,5}-[1-9][0-9]{0,8})$",
              "title": "Idea",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdeaDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not signed in (unauthorized)"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not allowed (forbidden, csrf_failed, or a specific code)"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found, or not visible to you (not_found)"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflicts with the current state (see code)"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            },
            "description": "Validation Failed"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Close evaluation",
        "tags": [
          "ideas"
        ]
      }
    },
    "/api/v1/ideas/{idea}/evaluation/due-date": {
      "put": {
        "description": "Owner or project admin. The body is the complete new value: due_at null means no due date. 409 evaluation_closed unless evaluation is open.",
        "operationId": "set_evaluation_due_date",
        "parameters": [
          {
            "description": "The idea: its id (UUID) or its key such as \"CUST-12\" (any case).",
            "in": "path",
            "name": "idea",
            "required": true,
            "schema": {
              "description": "The idea: its id (UUID) or its key such as \"CUST-12\" (any case).",
              "maxLength": 36,
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}|[A-Za-z][A-Za-z0-9]{1,5}-[1-9][0-9]{0,8})$",
              "title": "Idea",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluationDueDate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdeaDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not signed in (unauthorized)"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not allowed (forbidden, csrf_failed, or a specific code)"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found, or not visible to you (not_found)"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflicts with the current state (see code)"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            },
            "description": "Validation Failed"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Set or clear the evaluation due date",
        "tags": [
          "ideas"
        ]
      }
    },
    "/api/v1/ideas/{idea}/evaluation/reopen": {
      "post": {
        "description": "Owner or project admin. Idempotent. 409 idea_closed.",
        "operationId": "reopen_evaluation",
        "parameters": [
          {
            "description": "The idea: its id (UUID) or its key such as \"CUST-12\" (any case).",
            "in": "path",
            "name": "idea",
            "required": true,
            "schema": {
              "description": "The idea: its id (UUID) or its key such as \"CUST-12\" (any case).",
              "maxLength": 36,
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}|[A-Za-z][A-Za-z0-9]{1,5}-[1-9][0-9]{0,8})$",
              "title": "Idea",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdeaDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not signed in (unauthorized)"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not allowed (forbidden, csrf_failed, or a specific code)"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found, or not visible to you (not_found)"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflicts with the current state (see code)"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            },
            "description": "Validation Failed"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Reopen evaluation",
        "tags": [
          "ideas"
        ]
      }
    },
...
  }
}